+++ clothlogix/src/main/java/com/fabless/clothlogix/model/entities/TagliaEntity.java
@@ -13,13 +13,14 @@ public class TagliaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column (name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name =  "descrizione")
     private String descrizione;
 
     public TagliaEntity(){}
-    public TagliaEntity(Integer id, String descrizione) {
+
+    public TagliaEntity(Long id, String descrizione) {
         this.id = id;
         this.descrizione = descrizione;
     }
